Tweak one way message handling and call lifecycle events
The wrapped response object (`wres`) will now be consistent with the
implementation-specific behavior of one-way messages: stateless channels
actually did receive a respones, so it will now be available there
(along with a callback for client calls). The behavior is unchanged for
stateful channels (since no response was actually sent then).

Similarly, the "response" half of each call's "lifecycle events" will
only be emitted when a response is sent/received.